Automatic refresh only if buffer has changed
Now the function that refreshes the buffer on the idle timer only does
the actual refresh if the value of (buffer-hash) is different than the
one stored at the previous run.  This fixes the annoying behavior of
the buffer being constantly modified by 'numbex-refresh' even when
there is nothing to do.